% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-solutions.R
\docType{data}
\name{aoc_solutions}
\alias{aoc_solutions}
\title{Advent of Code 2021 solutions}
\format{
An object of class \code{list} of length 50.
}
\usage{
aoc_solutions
}
\description{
Advent of Code 2021 solutions
}
\keyword{datasets}
