% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day09.R
\name{basin_size}
\alias{basin_size}
\title{find the size of a basin}
\usage{
basin_size(file, xpos, ypos)
}
\arguments{
\item{file}{input file}

\item{xpos}{x position of minimum}

\item{ypos}{y position of minimum}
}
\description{
find the size of a basin
}
