% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{completion_score}
\alias{completion_score}
\title{Compute the completion score (part 2)}
\usage{
completion_score(file)
}
\arguments{
\item{file}{input file}
}
\description{
Compute the completion score (part 2)
}
\examples{
completion_score(system.file("extdata", "day10.txt", package = "adventofcode21"))

}
