% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{convert_binary_to_decimal}
\alias{convert_binary_to_decimal}
\title{Convert a vector containing binary bits to a decimal integer}
\usage{
convert_binary_to_decimal(x)
}
\arguments{
\item{x}{A vector containing the binary bits}
}
\description{
Convert a vector containing binary bits to a decimal integer
}
