% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{corruption_score}
\alias{corruption_score}
\title{Compute the corruption score (part 1)}
\usage{
corruption_score(file)
}
\arguments{
\item{file}{input file}
}
\description{
Compute the corruption score (part 1)
}
\examples{
corruption_score(system.file("extdata", "day10.txt", package = "adventofcode21"))
}
