% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day11.R
\name{count_flashes}
\alias{count_flashes}
\title{part 1}
\usage{
count_flashes(file)
}
\arguments{
\item{file}{input file}
}
\description{
part 1
}
\examples{
\dontrun{
file <- system.file("extdata", "day11.txt", package = "adventofcode21")
count_flashes(file)}


}
