% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day01.R
\name{count_increased}
\alias{count_increased}
\title{Count number of increasing values in a sequence}
\usage{
count_increased(x)
}
\arguments{
\item{x}{a numeric sequence}
}
\description{
Count number of increasing values in a sequence
}
\examples{
f <- system.file("extdata", "day1.txt", package = "adventofcode21")
x <- read.delim(f, header = FALSE)[[1]]
count_increased(x) # 1451

}
