% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{day03}
\alias{day03}
\alias{f03a}
\alias{f03b}
\alias{example_data_03}
\title{Day 03: Binary Diagnostic}
\usage{
f03a(x)

f03b(x)

example_data_03(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f03a(x)} returns .... For Part Two,
\code{f03b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2021/day/3}{Binary Diagnostic}
}
\details{
\strong{Part One}

The submarine has been making some [odd creaking
noises]{title="Turns out oceans are heavy."}, so you ask it to produce a
diagnostic report just in case.

The diagnostic report (your puzzle input) consists of a list of binary
numbers which, when decoded properly, can tell you many useful things
about the conditions of the submarine. The first parameter to check is
the \emph{power consumption}.

You need to use the binary numbers in the diagnostic report to generate
two new binary numbers (called the \emph{gamma rate} and the \emph{epsilon rate}).
The power consumption can then be found by multiplying the gamma rate by
the epsilon rate.

Each bit in the gamma rate can be determined by finding the \emph{most common
bit in the corresponding position} of all numbers in the diagnostic
report. For example, given the following diagnostic report:\preformatted{00100
11110
10110
10111
10101
01111
00111
11100
10000
11001
00010
01010
}

Considering only the first bit of each number, there are five \code{0} bits
and seven \code{1} bits. Since the most common bit is \code{1}, the first bit of
the gamma rate is \code{1}.

The most common second bit of the numbers in the diagnostic report is
\code{0}, so the second bit of the gamma rate is \code{0}.

The most common value of the third, fourth, and fifth bits are \code{1}, \code{1},
and \code{0}, respectively, and so the final three bits of the gamma rate are
\code{110}.

So, the gamma rate is the binary number \code{10110}, or \code{22} in decimal.

The epsilon rate is calculated in a similar way; rather than use the
most common bit, the least common bit from each position is used. So,
the epsilon rate is \code{01001}, or \code{9} in decimal. Multiplying the gamma
rate (\code{22}) by the epsilon rate (\code{9}) produces the power consumption,
\code{198}.

Use the binary numbers in your diagnostic report to calculate the gamma
rate and epsilon rate, then multiply them together. \emph{What is the power
consumption of the submarine?} (Be sure to represent your answer in
decimal, not binary.)

\strong{Part Two}

Next, you should verify the \emph{life support rating}, which can be
determined by multiplying the \emph{oxygen generator rating} by the \emph{CO2
scrubber rating}.

Both the oxygen generator rating and the CO2 scrubber rating are values
that can be found in your diagnostic report - finding them is the tricky
part. Both values are located using a similar process that involves
filtering out values until only one remains. Before searching for either
rating value, start with the full list of binary numbers from your
diagnostic report and \emph{consider just the first bit} of those numbers.
Then:
\itemize{
\item Keep only numbers selected by the \emph{bit criteria} for the type of
rating value for which you are searching. Discard numbers which do
not match the bit criteria.
\item If you only have one number left, stop; this is the rating value for
which you are searching.
\item Otherwise, repeat the process, considering the next bit to the
right.
}

The \emph{bit criteria} depends on which type of rating value you want to
find:
\itemize{
\item To find \emph{oxygen generator rating}, determine the \emph{most common} value
(\code{0} or \code{1}) in the current bit position, and keep only numbers with
that bit in that position. If \code{0} and \code{1} are equally common, keep
values with a \code{1} in the position being considered.
\item To find \emph{CO2 scrubber rating}, determine the \emph{least common} value
(\code{0} or \code{1}) in the current bit position, and keep only numbers with
that bit in that position. If \code{0} and \code{1} are equally common, keep
values with a \code{0} in the position being considered.
}

For example, to determine the \emph{oxygen generator rating} value using the
same example diagnostic report from above:
\itemize{
\item Start with all 12 numbers and consider only the first bit of each
number. There are more \code{1} bits (7) than \code{0} bits (5), so keep only
the 7 numbers with a \code{1} in the first position: \code{11110}, \code{10110},
\code{10111}, \code{10101}, \code{11100}, \code{10000}, and \code{11001}.
\item Then, consider the second bit of the 7 remaining numbers: there are
more \code{0} bits (4) than \code{1} bits (3), so keep only the 4 numbers with
a \code{0} in the second position: \code{10110}, \code{10111}, \code{10101}, and
\code{10000}.
\item In the third position, three of the four numbers have a \code{1}, so keep
those three: \code{10110}, \code{10111}, and \code{10101}.
\item In the fourth position, two of the three numbers have a \code{1}, so keep
those two: \code{10110} and \code{10111}.
\item In the fifth position, there are an equal number of \code{0} bits and \code{1}
bits (one each). So, to find the \emph{oxygen generator rating}, keep the
number with a \code{1} in that position: \code{10111}.
\item As there is only one number left, stop; the \emph{oxygen generator
rating} is \code{10111}, or \code{23} in decimal.
}

Then, to determine the \emph{CO2 scrubber rating} value from the same example
above:
\itemize{
\item Start again with all 12 numbers and consider only the first bit of
each number. There are fewer \code{0} bits (5) than \code{1} bits (7), so keep
only the 5 numbers with a \code{0} in the first position: \code{00100},
\code{01111}, \code{00111}, \code{00010}, and \code{01010}.
\item Then, consider the second bit of the 5 remaining numbers: there are
fewer \code{1} bits (2) than \code{0} bits (3), so keep only the 2 numbers
with a \code{1} in the second position: \code{01111} and \code{01010}.
\item In the third position, there are an equal number of \code{0} bits and \code{1}
bits (one each). So, to find the \emph{CO2 scrubber rating}, keep the
number with a \code{0} in that position: \code{01010}.
\item As there is only one number left, stop; the \emph{CO2 scrubber rating} is
\code{01010}, or \code{10} in decimal.
}

Finally, to find the life support rating, multiply the oxygen generator
rating (\code{23}) by the CO2 scrubber rating (\code{10}) to get \code{230}.

Use the binary numbers in your diagnostic report to calculate the oxygen
generator rating and CO2 scrubber rating, then multiply them together.
\emph{What is the life support rating of the submarine?} (Be sure to
represent your answer in decimal, not binary.)
}
\examples{
f03a(example_data_03())
f03b(example_data_03())
}
