% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05}
\alias{day05}
\alias{f05a}
\alias{f05b}
\alias{example_data_05}
\title{Day 05: Hydrothermal Venture}
\usage{
f05a(x)

f05b(x)

example_data_05(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f05a(x)} returns .... For Part Two,
\code{f05b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2021/day/5}{Hydrothermal Venture}
}
\details{
\strong{Part One}

You come across a field of \href{https://en.wikipedia.org/wiki/Hydrothermal_vent}{hydrothermal vents} on the ocean
floor! These vents constantly produce large, opaque clouds, so it would
be best to avoid them if possible.

They tend to form in \emph{lines}; the submarine helpfully produces a list of
nearby [lines of vents]{title="Maybe they're Bresenham vents."} (your
puzzle input) for you to review. For example:\preformatted{0,9 -> 5,9
8,0 -> 0,8
9,4 -> 3,4
2,2 -> 2,1
7,0 -> 7,4
6,4 -> 2,0
0,9 -> 2,9
3,4 -> 1,4
0,0 -> 8,8
5,5 -> 8,2
}

Each line of vents is given as a line segment in the format
\verb{x1,y1 -> x2,y2} where \code{x1},\code{y1} are the coordinates of one end the line
segment and \code{x2},\code{y2} are the coordinates of the other end. These line
segments include the points at both ends. In other words:
\itemize{
\item An entry like \verb{1,1 -> 1,3} covers points \verb{1,1}, \verb{1,2}, and \verb{1,3}.
\item An entry like \verb{9,7 -> 7,7} covers points \verb{9,7}, \verb{8,7}, and \verb{7,7}.
}

For now, \emph{only consider horizontal and vertical lines}: lines where
either \code{x1 = x2} or \code{y1 = y2}.

So, the horizontal and vertical lines from the above list would produce
the following diagram:\preformatted{.......1..
..1....1..
..1....1..
.......1..
.112111211
..........
..........
..........
..........
222111....
}

In this diagram, the top left corner is \verb{0,0} and the bottom right
corner is \verb{9,9}. Each position is shown as \emph{the number of lines which
cover that point} or \code{.} if no line covers that point. The top-left pair
of \code{1}s, for example, comes from \verb{2,2 -> 2,1}; the very bottom row is
formed by the overlapping lines \verb{0,9 -> 5,9} and \verb{0,9 -> 2,9}.

To avoid the most dangerous areas, you need to determine \emph{the number of
points where at least two lines overlap}. In the above example, this is
anywhere in the diagram with a \code{2} or larger - a total of \code{5} points.

Consider only horizontal and vertical lines. \emph{At how many points do at
least two lines overlap?}

\strong{Part Two}

Unfortunately, considering only horizontal and vertical lines doesn\'t
give you the full picture; you need to also consider \emph{diagonal lines}.

Because of the limits of the hydrothermal vent mapping system, the lines
in your list will only ever be horizontal, vertical, or a diagonal line
at exactly 45 degrees. In other words:
\itemize{
\item An entry like \verb{1,1 -> 3,3} covers points \verb{1,1}, \verb{2,2}, and \verb{3,3}.
\item An entry like \verb{9,7 -> 7,9} covers points \verb{9,7}, \verb{8,8}, and \verb{7,9}.
}

Considering all lines from the above example would now produce the
following diagram:\preformatted{1.1....11.
.111...2..
..2.1.111.
...1.2.2..
.112313211
...1.2....
..1...1...
.1.....1..
1.......1.
222111....
}

You still need to determine \emph{the number of points where at least two
lines overlap}. In the above example, this is still anywhere in the
diagram with a \code{2} or larger - now a total of \code{12} points.

Consider all of the lines. \emph{At how many points do at least two lines
overlap?}
}
\examples{
f05a(example_data_05())
f05b(example_data_05())
}
