% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day08.R
\name{day08}
\alias{day08}
\alias{f08a}
\alias{f08b}
\alias{example_data_08}
\title{Day 08: Seven Segment Search}
\usage{
f08a(x)

f08b(x)

example_data_08(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f08a(x)} returns .... For Part Two,
\code{f08b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2021/day/8}{Seven Segment Search}
}
\details{
\strong{Part One}

You barely reach the safety of the cave when the whale smashes into the
cave mouth, collapsing it. Sensors indicate another exit to this cave at
a much greater depth, so you have no choice but to press on.

As your submarine slowly makes its way through the cave system, you
notice that the four-digit \href{https://en.wikipedia.org/wiki/Seven-segment_display}{seven-segment displays} in your
submarine are malfunctioning; [they must have been
damaged]{title="Yes, just the four-digit seven-segment ones. Whole batch must have been faulty."}
during the escape. You\'ll be in a lot of trouble without them, so
you\'d better figure out what\'s wrong.

Each digit of a seven-segment display is rendered by turning on or off
any of seven segments named \code{a} through \code{g}:\preformatted{  0:      1:      2:      3:      4:
 aaaa    ....    aaaa    aaaa    ....
b    c  .    c  .    c  .    c  b    c
b    c  .    c  .    c  .    c  b    c
 ....    ....    dddd    dddd    dddd
e    f  .    f  e    .  .    f  .    f
e    f  .    f  e    .  .    f  .    f
 gggg    ....    gggg    gggg    ....

  5:      6:      7:      8:      9:
 aaaa    aaaa    aaaa    aaaa    aaaa
b    .  b    .  .    c  b    c  b    c
b    .  b    .  .    c  b    c  b    c
 dddd    dddd    ....    dddd    dddd
.    f  e    f  .    f  e    f  .    f
.    f  e    f  .    f  e    f  .    f
 gggg    gggg    ....    gggg    gggg
}

So, to render a \code{1}, only segments \code{c} and \code{f} would be turned on; the
rest would be off. To render a \code{7}, only segments \code{a}, \code{c}, and \code{f}
would be turned on.

The problem is that the signals which control the segments have been
mixed up on each display. The submarine is still trying to display
numbers by producing output on signal wires \code{a} through \code{g}, but those
wires are connected to segments \emph{randomly}. Worse, the wire/segment
connections are mixed up separately for each four-digit display! (All of
the digits \emph{within} a display use the same connections, though.)

So, you might know that only signal wires \code{b} and \code{g} are turned on, but
that doesn\'t mean \emph{segments} \code{b} and \code{g} are turned on: the only digit
that uses two segments is \code{1}, so it must mean segments \code{c} and \code{f} are
meant to be on. With just that information, you still can\'t tell which
wire (\code{b}/\code{g}) goes to which segment (\code{c}/\code{f}). For that, you\'ll need
to collect more information.

For each display, you watch the changing signals for a while, make a
note of \emph{all ten unique signal patterns} you see, and then write down a
single \emph{four digit output value} (your puzzle input). Using the signal
patterns, you should be able to work out which pattern corresponds to
which digit.

For example, here is what you might see in a single entry in your notes:\preformatted{acedgfb cdfbe gcdfa fbcad dab cefabd cdfgeb eafb cagedb ab |
cdfeb fcadb cdfeb cdbaf
}

(The entry is wrapped here to two lines so it fits; in your notes, it
will all be on a single line.)

Each entry consists of ten \emph{unique signal patterns}, a \code{|} delimiter,
and finally the \emph{four digit output value}. Within an entry, the same
wire/segment connections are used (but you don\'t know what the
connections actually are). The unique signal patterns correspond to the
ten different ways the submarine tries to render a digit using the
current wire/segment connections. Because \code{7} is the only digit that
uses three segments, \code{dab} in the above example means that to render a
\code{7}, signal lines \code{d}, \code{a}, and \code{b} are on. Because \code{4} is the only
digit that uses four segments, \code{eafb} means that to render a \code{4}, signal
lines \code{e}, \code{a}, \code{f}, and \code{b} are on.

Using this information, you should be able to work out which combination
of signal wires corresponds to each of the ten digits. Then, you can
decode the four digit output value. Unfortunately, in the above example,
all of the digits in the output value (\verb{cdfeb fcadb cdfeb cdbaf}) use
five segments and are more difficult to deduce.

For now, \emph{focus on the easy digits}. Consider this larger example:\preformatted{be cfbegad cbdgef fgaecd cgeb fdcge agebfd fecdb fabcd edb |
fdgacbe cefdb cefbgd gcbe
edbfga begcd cbg gc gcadebf fbgde acbgfd abcde gfcbed gfec |
fcgedb cgb dgebacf gc
fgaebd cg bdaec gdafb agbcfd gdcbef bgcad gfac gcb cdgabef |
cg cg fdcagb cbg
fbegcd cbd adcefb dageb afcb bc aefdc ecdab fgdeca fcdbega |
efabcd cedba gadfec cb
aecbfdg fbg gf bafeg dbefa fcge gcbea fcaegb dgceab fcbdga |
gecf egdcabf bgf bfgea
fgeab ca afcebg bdacfeg cfaedg gcfdb baec bfadeg bafgc acf |
gebdcfa ecba ca fadegcb
dbcfg fgd bdegcaf fgec aegbdf ecdfab fbedc dacgb gdcebf gf |
cefg dcbef fcge gbcadfe
bdfegc cbegaf gecbf dfcage bdacg ed bedf ced adcbefg gebcd |
ed bcgafe cdgba cbgef
egadfb cdbfeg cegd fecab cgb gbdefca cg fgcdab egfdb bfceg |
gbdfcae bgc cg cgb
gcafb gcf dcaebfg ecagb gf abcdeg gaef cafbge fdbac fegbdc |
fgae cfgab fg bagce
}

Because the digits \code{1}, \code{4}, \code{7}, and \code{8} each use a unique number of
segments, you should be able to tell which combinations of signals
correspond to those digits. Counting \emph{only digits in the output values}
(the part after \code{|} on each line), in the above example, there are \code{26}
instances of digits that use a unique number of segments (highlighted
above).

\emph{In the output values, how many times do digits \code{1}, \code{4}, \code{7}, or \code{8}
appear?}

\strong{Part Two}

Through a little deduction, you should now be able to determine the
remaining digits. Consider again the first example above:\preformatted{acedgfb cdfbe gcdfa fbcad dab cefabd cdfgeb eafb cagedb ab |
cdfeb fcadb cdfeb cdbaf
}

After some careful analysis, the mapping between signal wires and
segments only make sense in the following configuration:\preformatted{ dddd
e    a
e    a
 ffff
g    b
g    b
 cccc
}

So, the unique signal patterns would correspond to the following digits:
\itemize{
\item \code{acedgfb}: \code{8}
\item \code{cdfbe}: \code{5}
\item \code{gcdfa}: \code{2}
\item \code{fbcad}: \code{3}
\item \code{dab}: \code{7}
\item \code{cefabd}: \code{9}
\item \code{cdfgeb}: \code{6}
\item \code{eafb}: \code{4}
\item \code{cagedb}: \code{0}
\item \code{ab}: \code{1}
}

Then, the four digits of the output value can be decoded:
\itemize{
\item \code{cdfeb}: \code{5}
\item \code{fcadb}: \code{3}
\item \code{cdfeb}: \code{5}
\item \code{cdbaf}: \code{3}
}

Therefore, the output value for this entry is \code{5353}.

Following this same process for each entry in the second, larger example
above, the output value of each entry can be determined:
\itemize{
\item \verb{fdgacbe cefdb cefbgd gcbe}: \code{8394}
\item \verb{fcgedb cgb dgebacf gc}: \code{9781}
\item \verb{cg cg fdcagb cbg}: \code{1197}
\item \verb{efabcd cedba gadfec cb}: \code{9361}
\item \verb{gecf egdcabf bgf bfgea}: \code{4873}
\item \verb{gebdcfa ecba ca fadegcb}: \code{8418}
\item \verb{cefg dcbef fcge gbcadfe}: \code{4548}
\item \verb{ed bcgafe cdgba cbgef}: \code{1625}
\item \verb{gbdfcae bgc cg cgb}: \code{8717}
\item \verb{fgae cfgab fg bagce}: \code{4315}
}

Adding all of the output values in this larger example produces \code{61229}.

For each entry, determine all of the wire/segment connections and decode
the four-digit output values. \emph{What do you get if you add up all of the
output values?}
}
\examples{
f08a(example_data_08())
f08b(example_data_08())
}
