% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day09.R
\name{day09}
\alias{day09}
\alias{f09a}
\alias{f09b}
\alias{example_data_09}
\title{Day 09: Smoke Basin}
\usage{
f09a(x)

f09b(x)

example_data_09(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f09a(x)} returns .... For Part Two,
\code{f09b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2021/day/9}{Smoke Basin}
}
\details{
\strong{Part One}

These caves seem to be \href{https://en.wikipedia.org/wiki/Lava_tube}{lava tubes}. Parts are even still
volcanically active; small hydrothermal vents release smoke into the
caves that slowly [settles like
rain]{title="This was originally going to be a puzzle about watersheds, but we're already under water."}.

If you can model how the smoke flows through the caves, you might be
able to avoid it and be that much safer. The submarine generates a
heightmap of the floor of the nearby caves for you (your puzzle input).

Smoke flows to the lowest point of the area it\'s in. For example,
consider the following heightmap:\preformatted{2199943210
3987894921
9856789892
8767896789
9899965678
}

Each number corresponds to the height of a particular location, where
\code{9} is the highest and \code{0} is the lowest a location can be.

Your first goal is to find the \emph{low points} - the locations that are
lower than any of its adjacent locations. Most locations have four
adjacent locations (up, down, left, and right); locations on the edge or
corner of the map have three or two adjacent locations, respectively.
(Diagonal locations do not count as adjacent.)

In the above example, there are \emph{four} low points, all highlighted: two
are in the first row (a \code{1} and a \code{0}), one is in the third row (a \code{5}),
and one is in the bottom row (also a \code{5}). All other locations on the
heightmap have some lower adjacent location, and so are not low points.

The \emph{risk level} of a low point is \emph{1 plus its height}. In the above
example, the risk levels of the low points are \code{2}, \code{1}, \code{6}, and \code{6}.
The sum of the risk levels of all low points in the heightmap is
therefore \code{15}.

Find all of the low points on your heightmap. \emph{What is the sum of the
risk levels of all low points on your heightmap?}

\strong{Part Two}

Next, you need to find the largest basins so you know what areas are
most important to avoid.

A \emph{basin} is all locations that eventually flow downward to a single low
point. Therefore, every low point has a basin, although some basins are
very small. Locations of height \code{9} do not count as being in any basin,
and all other locations will always be part of exactly one basin.

The \emph{size} of a basin is the number of locations within the basin,
including the low point. The example above has four basins.

The top-left basin, size \code{3}:\preformatted{2199943210
3987894921
9856789892
8767896789
9899965678
}

The top-right basin, size \code{9}:\preformatted{2199943210
3987894921
9856789892
8767896789
9899965678
}

The middle basin, size \code{14}:\preformatted{2199943210
3987894921
9856789892
8767896789
9899965678
}

The bottom-right basin, size \code{9}:\preformatted{2199943210
3987894921
9856789892
8767896789
9899965678
}

Find the three largest basins and multiply their sizes together. In the
above example, this is \code{9 * 14 * 9 = 1134}.

\emph{What do you get if you multiply together the sizes of the three largest
basins?}
}
\examples{
f09a(example_data_09())
f09b(example_data_09())
}
