% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day12.R
\name{day12}
\alias{day12}
\alias{f12a_traverse_caves}
\alias{f12b}
\alias{example_data_12}
\title{Day 12: Passage Pathing}
\usage{
f12a_traverse_caves(x)

f12b(x)

example_data_12(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f12a(x)} returns .... For Part Two,
\code{f12b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2021/day/12}{Passage Pathing}
}
\details{
\strong{Part One}

With your [submarine\'s subterranean subsystems subsisting
suboptimally]{title="Sublime."}, the only way you\'re getting out of
this cave anytime soon is by finding a path yourself. Not just \emph{a} path
\itemize{
\item the only way to know if you\'ve found the \emph{best} path is to find \emph{all}
of them.
}

Fortunately, the sensors are still mostly working, and so you build a
rough map of the remaining caves (your puzzle input). For example:\preformatted{start-A
start-b
A-c
A-b
b-d
A-end
b-end
}

This is a list of how all of the caves are connected. You start in the
cave named \code{start}, and your destination is the cave named \code{end}. An
entry like \code{b-d} means that cave \code{b} is connected to cave \code{d} - that is,
you can move between them.

So, the above cave system looks roughly like this:\preformatted{    start
    /   \\
c--A-----b--d
    \\   /
     end
}

Your goal is to find the number of distinct \emph{paths} that start at
\code{start}, end at \code{end}, and don\'t visit small caves more than once.
There are two types of caves: \emph{big} caves (written in uppercase, like
\code{A}) and \emph{small} caves (written in lowercase, like \code{b}). It would be a
waste of time to visit any small cave more than once, but big caves are
large enough that it might be worth visiting them multiple times. So,
all paths you find should \emph{visit small caves at most once}, and can
\emph{visit big caves any number of times}.

Given these rules, there are \code{10} paths through this example cave
system:\preformatted{start,A,b,A,c,A,end
start,A,b,A,end
start,A,b,end
start,A,c,A,b,A,end
start,A,c,A,b,end
start,A,c,A,end
start,A,end
start,b,A,c,A,end
start,b,A,end
start,b,end
}

(Each line in the above list corresponds to a single path; the caves
visited by that path are listed in the order they are visited and
separated by commas.)

Note that in this cave system, cave \code{d} is never visited by any path: to
do so, cave \code{b} would need to be visited twice (once on the way to cave
\code{d} and a second time when returning from cave \code{d}), and since cave \code{b}
is small, this is not allowed.

Here is a slightly larger example:\preformatted{dc-end
HN-start
start-kj
dc-start
dc-HN
LN-dc
HN-end
kj-sa
kj-HN
kj-dc
}

The \code{19} paths through it are as follows:\preformatted{start,HN,dc,HN,end
start,HN,dc,HN,kj,HN,end
start,HN,dc,end
start,HN,dc,kj,HN,end
start,HN,end
start,HN,kj,HN,dc,HN,end
start,HN,kj,HN,dc,end
start,HN,kj,HN,end
start,HN,kj,dc,HN,end
start,HN,kj,dc,end
start,dc,HN,end
start,dc,HN,kj,HN,end
start,dc,end
start,dc,kj,HN,end
start,kj,HN,dc,HN,end
start,kj,HN,dc,end
start,kj,HN,end
start,kj,dc,HN,end
start,kj,dc,end
}

Finally, this even larger example has \code{226} paths through it:\preformatted{fs-end
he-DX
fs-he
start-DX
pj-DX
end-zg
zg-sl
zg-pj
pj-he
RW-he
fs-DX
pj-RW
zg-RW
start-pj
he-WI
zg-he
pj-fs
start-RW
}

\emph{How many paths through this cave system are there that visit small
caves at most once?}

\strong{Part Two}

\emph{(Use have to manually add this yourself.)}

\emph{(Try using \code{convert_clipboard_html_to_roxygen_md()})}
}
\examples{
f12a_traverse_caves(example_data_12())
f12b()
}
