% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day13.R
\name{day13}
\alias{day13}
\alias{f13a}
\alias{f13b}
\alias{example_data_13}
\title{Day 13: Transparent Origami}
\usage{
f13a(x)

f13b(x)

example_data_13(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f13a(x)} returns .... For Part Two,
\code{f13b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2021/day/13}{Transparent Origami}
}
\details{
\strong{Part One}

You reach another volcanically active part of the cave. It would be nice
if you could do some kind of thermal imaging so you could tell ahead of
time which caves are too hot to safely enter.

Fortunately, the submarine seems to be equipped with a thermal camera!
When you activate it, you are greeted with:\preformatted{Congratulations on your purchase! To activate this infrared thermal imaging
camera system, please enter the code found on page 1 of the manual.
}

Apparently, the Elves have never used this feature. To your surprise,
you manage to find the manual; as you go to open it, page 1 falls out.
It\'s a large sheet of \href{https://en.wikipedia.org/wiki/Transparency_(projection)}{transparent paper}! The
transparent paper is marked with random dots and includes instructions
on how to fold it up (your puzzle input). For example:\preformatted{6,10
0,14
9,10
0,3
10,4
4,11
6,0
6,12
4,1
0,13
10,12
3,4
3,0
8,4
1,10
2,14
8,10
9,0

fold along y=7
fold along x=5
}

The first section is a list of dots on the transparent paper. \verb{0,0}
represents the top-left coordinate. The first value, \code{x}, increases to
the right. The second value, \code{y}, increases downward. So, the coordinate
\verb{3,0} is to the right of \verb{0,0}, and the coordinate \verb{0,7} is below \verb{0,0}.
The coordinates in this example form the following pattern, where \verb{#} is
a dot on the paper and \code{.} is an empty, unmarked position:\preformatted{...#..#..#.
....#......
...........
#..........
...#....#.#
...........
...........
...........
...........
...........
.#....#.##.
....#......
......#...#
#..........
#.#........
}

Then, there is a list of \emph{fold instructions}. Each instruction indicates
a line on the transparent paper and wants you to fold the paper \emph{up}
(for horizontal \code{y=...} lines) or \emph{left} (for vertical \code{x=...} lines).
In this example, the first fold instruction is \verb{fold along y=7}, which
designates the line formed by all of the positions where \code{y} is \code{7}
(marked here with \code{-}):\preformatted{...#..#..#.
....#......
...........
#..........
...#....#.#
...........
...........
-----------
...........
...........
.#....#.##.
....#......
......#...#
#..........
#.#........
}

Because this is a horizontal line, fold the bottom half \emph{up}. Some of
the dots might end up overlapping after the fold is complete, but dots
will never appear exactly on a fold line. The result of doing this fold
looks like this:\preformatted{#.##..#..#.
#...#......
......#...#
#...#......
.#.#..#.###
...........
...........
}

Now, only \code{17} dots are visible.

Notice, for example, the two dots in the bottom left corner before the
transparent paper is folded; after the fold is complete, those dots
appear in the top left corner (at \verb{0,0} and \verb{0,1}). Because the paper is
transparent, the dot just below them in the result (at \verb{0,3}) remains
visible, as it can be seen through the transparent paper.

Also notice that some dots can end up \emph{overlapping}; in this case, the
dots merge together and become a single dot.

The second fold instruction is \verb{fold along x=5}, which indicates this
line:\preformatted{#.##.|#..#.
#...#|.....
.....|#...#
#...#|.....
.#.#.|#.###
.....|.....
.....|.....
}

Because this is a vertical line, fold \emph{left}:\preformatted{#####
#...#
#...#
#...#
#####
.....
.....
}

The instructions made a square!

The transparent paper is pretty big, so for now, focus on just
completing the first fold. After the first fold in the example above,
\code{17} dots are visible - dots that end up overlapping after the fold is
completed count as a single dot.

\emph{How many dots are visible after completing just the first fold
instruction on your transparent paper?}

\strong{Part Two}

\emph{(Use have to manually add this yourself.)}

\emph{(Try using \code{convert_clipboard_html_to_roxygen_md()})}
}
\examples{
f13a(example_data_13())
f13b(example_data_13())
}
