% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day15.R
\name{day15}
\alias{day15}
\alias{f15a_dijkstra}
\alias{f15b}
\alias{example_data_15}
\title{Day 15: Chiton}
\usage{
f15a_dijkstra(x)

f15b(x)

example_data_15(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f15a(x)} returns .... For Part Two,
\code{f15b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2021/day/15}{Chiton}
}
\details{
\strong{Part One}

You\'ve almost reached the exit of the cave, but the walls are getting
closer together. Your submarine can barely still fit, though; the main
problem is that the walls of the cave are covered in
\href{https://en.wikipedia.org/wiki/Chiton}{chitons}, and it would be best
not to bump any of them.

The cavern is large, but has a very low ceiling, restricting your motion
to two dimensions. The shape of the cavern resembles a square; a quick
scan of chiton density produces a map of \emph{risk level} throughout the
cave (your puzzle input). For example:\preformatted{1163751742
1381373672
2136511328
3694931569
7463417111
1319128137
1359912421
3125421639
1293138521
2311944581
}

You start in the top left position, your destination is the bottom right
position, and you [cannot move
diagonally]{title="Can't go diagonal until we can repair the caterpillar unit. Could be the liquid helium or the superconductors."}.
The number at each position is its \emph{risk level}; to determine the total
risk of an entire path, add up the risk levels of each position you
\emph{enter} (that is, don\'t count the risk level of your starting position
unless you enter it; leaving it adds no risk to your total).

Your goal is to find a path with the \emph{lowest total risk}. In this
example, a path with the lowest total risk is highlighted here:\preformatted{1163751742
1381373672
2136511328
3694931569
7463417111
1319128137
1359912421
3125421639
1293138521
2311944581
}

The total risk of this path is \code{40} (the starting position is never
entered, so its risk is not counted).

\emph{What is the lowest total risk of any path from the top left to the
bottom right?}

\strong{Part Two}

\emph{(Use have to manually add this yourself.)}

\emph{(Try using \code{convert_clipboard_html_to_roxygen_md()})}
}
\examples{
f15a(example_data_15())
f15b()
}
