% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day17.R
\name{day17}
\alias{day17}
\alias{f17a}
\alias{f17b}
\alias{example_data_17}
\title{Day 17: Trick Shot}
\usage{
f17a(x)

f17b(x, part1 = TRUE)

example_data_17(example = 1)
}
\arguments{
\item{x}{some data}

\item{part1}{set to \code{TRUE} if solving part 1, \code{FALSE} otherwise}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f17a(x)} returns .... For Part Two,
\code{f17b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2021/day/17}{Trick Shot}
}
\details{
\strong{Part One}

You finally decode the Elves\' message. \code{HI}, the message says. You
continue searching for the sleigh keys.

Ahead of you is what appears to be a large \href{https://en.wikipedia.org/wiki/Oceanic_trench}{ocean trench}. Could the keys
have fallen into it? You\'d better send a probe to investigate.

The probe launcher on your submarine can fire the probe with any
\href{https://en.wikipedia.org/wiki/Integer}{integer} velocity in the \code{x}
(forward) and \code{y} (upward, or downward if negative) directions. For
example, an initial \verb{x,y} velocity like \verb{0,10} would fire the probe
straight up, while an initial velocity like \verb{10,-1} would fire the probe
forward at a slight downward angle.

The probe\'s \verb{x,y} position starts at \verb{0,0}. Then, it will follow some
trajectory by moving in \emph{steps}. On each step, these changes occur in
the following order:
\itemize{
\item The probe\'s \code{x} position increases by its \code{x} velocity.
\item The probe\'s \code{y} position increases by its \code{y} velocity.
\item Due to drag, the probe\'s \code{x} velocity changes by \code{1} toward the
value \code{0}; that is, it decreases by \code{1} if it is greater than \code{0},
increases by \code{1} if it is less than \code{0}, or does not change if it is
already \code{0}.
\item Due to gravity, the probe\'s \code{y} velocity decreases by \code{1}.
}

For the probe to successfully make it into the trench, the probe must be
on some trajectory that causes it to be within a \emph{target area} after any
step. The submarine computer has already calculated this target area
(your puzzle input). For example:\preformatted{target area: x=20..30, y=-10..-5
}

This target area means that you need to find initial \verb{x,y} velocity
values such that after any step, the probe\'s \code{x} position is at least
\code{20} and at most \code{30}, \emph{and} the probe\'s \code{y} position is at least \code{-10}
and at most \code{-5}.

Given this target area, one initial velocity that causes the probe to be
within the target area after any step is \verb{7,2}:\preformatted{.............#....#............
.......#..............#........
...............................
S........................#.....
...............................
...............................
...........................#...
...............................
....................TTTTTTTTTTT
....................TTTTTTTTTTT
....................TTTTTTTT#TT
....................TTTTTTTTTTT
....................TTTTTTTTTTT
....................TTTTTTTTTTT
}

In this diagram, \code{S} is the probe\'s initial position, \verb{0,0}. The \code{x}
coordinate increases to the right, and the \code{y} coordinate increases
upward. In the bottom right, positions that are within the target area
are shown as \code{T}. After each step (until the target area is reached),
the position of the probe is marked with \verb{#}. (The bottom-right \verb{#} is
both a position the probe reaches and a position in the target area.)

Another initial velocity that causes the probe to be within the target
area after any step is \verb{6,3}:\preformatted{...............#..#............
...........#........#..........
...............................
......#..............#.........
...............................
...............................
S....................#.........
...............................
...............................
...............................
.....................#.........
....................TTTTTTTTTTT
....................TTTTTTTTTTT
....................TTTTTTTTTTT
....................TTTTTTTTTTT
....................T#TTTTTTTTT
....................TTTTTTTTTTT
}

Another one is \verb{9,0}:\preformatted{S........#.....................
.................#.............
...............................
........................#......
...............................
....................TTTTTTTTTTT
....................TTTTTTTTTT#
....................TTTTTTTTTTT
....................TTTTTTTTTTT
....................TTTTTTTTTTT
....................TTTTTTTTTTT
}

One initial velocity that \emph{doesn\'t} cause the probe to be within the
target area after any step is \verb{17,-4}:\preformatted{S..............................................................
...............................................................
...............................................................
...............................................................
.................#.............................................
....................TTTTTTTTTTT................................
....................TTTTTTTTTTT................................
....................TTTTTTTTTTT................................
....................TTTTTTTTTTT................................
....................TTTTTTTTTTT..#.............................
....................TTTTTTTTTTT................................
...............................................................
...............................................................
...............................................................
...............................................................
................................................#..............
...............................................................
...............................................................
...............................................................
...............................................................
...............................................................
...............................................................
..............................................................#
}

The probe appears to pass through the target area, but is never within
it after any step. Instead, it continues down and to the right - only
the first few steps are shown.

If you\'re going to fire a highly scientific probe out of a super cool
probe launcher, you might as well do it with \emph{style}. How high can you
make the probe go while still reaching the target area?

In the above example, using an initial velocity of \verb{6,9} is the best you
can do, causing the probe to reach a maximum \code{y} position of \code{45}. (Any
higher initial \code{y} velocity causes the probe to overshoot the target
area entirely.)

Find the initial velocity that causes the probe to reach the highest \code{y}
position and still eventually be within the target area after any step.
\emph{What is the highest \code{y} position it reaches on this trajectory?}

\strong{Part Two}

\emph{(Use have to manually add this yourself.)}

\emph{(Try using \code{convert_clipboard_html_to_roxygen_md()})}
}
\examples{
max(f17a(example_data_17()))
length(f17b(example_data_17()))
}
