% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day18.R
\name{day18}
\alias{day18}
\alias{f18a_snailfish_homework}
\alias{f18b_snailfish_homework2}
\alias{example_data_18}
\title{Day 18: Snailfish}
\usage{
f18a_snailfish_homework(x)

f18b_snailfish_homework2(x)

example_data_18(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f18a(x)} returns the magnitude of the sum of snailfish
numbers. For Part Two, \code{f18b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2021/day/18}{Snailfish}
}
\details{
\strong{Part One}

You descend into the ocean trench and encounter some
\href{https://en.wikipedia.org/wiki/Snailfish}{snailfish}. They say they saw
the sleigh keys! They\'ll even tell you which direction the keys went if
you help one of the smaller snailfish with his
\emph{[math]{title="Or 'maths', if you have more than one."} homework}.

Snailfish numbers aren\'t like regular numbers. Instead, every snailfish
number is a \emph{pair} - an ordered list of two elements. Each element of
the pair can be either a regular number or another pair.

Pairs are written as \verb{[x,y]}, where \code{x} and \code{y} are the elements within
the pair. Here are some example snailfish numbers, one snailfish number
per line:\preformatted{[1,2]
[[1,2],3]
[9,[8,7]]
[[1,9],[8,5]]
[[[[1,2],[3,4]],[[5,6],[7,8]]],9]
[[[9,[3,8]],[[0,9],6]],[[[3,7],[4,9]],3]]
[[[[1,3],[5,3]],[[1,3],[8,7]]],[[[4,9],[6,9]],[[8,2],[7,3]]]]
}

This snailfish homework is about \emph{addition}. To add two snailfish
numbers, form a pair from the left and right parameters of the addition
operator. For example, \verb{[1,2]} + \verb{[[3,4],5]} becomes
\verb{[[1,2],[[3,4],5]]}.

There\'s only one problem: \emph{snailfish numbers must always be reduced},
and the process of adding two snailfish numbers can result in snailfish
numbers that need to be reduced.

To \emph{reduce a snailfish number}, you must repeatedly do the first action
in this list that applies to the snailfish number:
\itemize{
\item If any pair is \emph{nested inside four pairs}, the leftmost such pair
\emph{explodes}.
\item If any regular number is \emph{10 or greater}, the leftmost such regular
number \emph{splits}.
}

Once no action in the above list applies, the snailfish number is
reduced.

During reduction, at most one action applies, after which the process
returns to the top of the list of actions. For example, if \emph{split}
produces a pair that meets the \emph{explode} criteria, that pair \emph{explodes}
before other \emph{splits} occur.

To \emph{explode} a pair, the pair\'s left value is added to the first
regular number to the left of the exploding pair (if any), and the
pair\'s right value is added to the first regular number to the right of
the exploding pair (if any). Exploding pairs will always consist of two
regular numbers. Then, the entire exploding pair is replaced with the
regular number \code{0}.

Here are some examples of a single explode action:
\itemize{
\item \verb{[[[[[9,8],1],2],3],4]} becomes \verb{[[[[0,9],2],3],4]} (the \code{9} has no
regular number to its left, so it is not added to any regular
number).
\item \verb{[7,[6,[5,[4,[3,2]]]]]} becomes \verb{[7,[6,[5,[7,0]]]]} (the \code{2} has no
regular number to its right, and so it is not added to any regular
number).
\item \verb{[[6,[5,[4,[3,2]]]],1]} becomes \verb{[[6,[5,[7,0]]],3]}.
\item \verb{[[3,[2,[1,[7,3]]]],[6,[5,[4,[3,2]]]]]} becomes
\verb{[[3,[2,[8,0]]],[9,[5,[4,[3,2]]]]]} (the pair \verb{[3,2]} is unaffected
because the pair \verb{[7,3]} is further to the left; \verb{[3,2]} would
explode on the next action).
\item \verb{[[3,[2,[8,0]]],[9,[5,[4,[3,2]]]]]} becomes
\verb{[[3,[2,[8,0]]],[9,[5,[7,0]]]]}.
}

To \emph{split} a regular number, replace it with a pair; the left element of
the pair should be the regular number divided by two and rounded \emph{down},
while the right element of the pair should be the regular number divided
by two and rounded \emph{up}. For example, \code{10} becomes \verb{[5,5]}, \code{11} becomes
\verb{[5,6]}, \code{12} becomes \verb{[6,6]}, and so on.

Here is the process of finding the reduced result of
\verb{[[[[4,3],4],4],[7,[[8,4],9]]]} + \verb{[1,1]}:\preformatted{after addition: [[[[[4,3],4],4],[7,[[8,4],9]]],[1,1]]
after explode:  [[[[0,7],4],[7,[[8,4],9]]],[1,1]]
after explode:  [[[[0,7],4],[15,[0,13]]],[1,1]]
after split:    [[[[0,7],4],[[7,8],[0,13]]],[1,1]]
after split:    [[[[0,7],4],[[7,8],[0,[6,7]]]],[1,1]]
after explode:  [[[[0,7],4],[[7,8],[6,0]]],[8,1]]
}

Once no reduce actions apply, the snailfish number that remains is the
actual result of the addition operation:
\verb{[[[[0,7],4],[[7,8],[6,0]]],[8,1]]}.

The homework assignment involves adding up a \emph{list of snailfish numbers}
(your puzzle input). The snailfish numbers are each listed on a separate
line. Add the first snailfish number and the second, then add that
result and the third, then add that result and the fourth, and so on
until all numbers in the list have been used once.

For example, the final sum of this list is
\verb{[[[[1,1],[2,2]],[3,3]],[4,4]]}:\preformatted{[1,1]
[2,2]
[3,3]
[4,4]
}

The final sum of this list is \verb{[[[[3,0],[5,3]],[4,4]],[5,5]]}:\preformatted{[1,1]
[2,2]
[3,3]
[4,4]
[5,5]
}

The final sum of this list is \verb{[[[[5,0],[7,4]],[5,5]],[6,6]]}:\preformatted{[1,1]
[2,2]
[3,3]
[4,4]
[5,5]
[6,6]
}

Here\'s a slightly larger example:\preformatted{[[[0,[4,5]],[0,0]],[[[4,5],[2,6]],[9,5]]]
[7,[[[3,7],[4,3]],[[6,3],[8,8]]]]
[[2,[[0,8],[3,4]]],[[[6,7],1],[7,[1,6]]]]
[[[[2,4],7],[6,[0,5]]],[[[6,8],[2,8]],[[2,1],[4,5]]]]
[7,[5,[[3,8],[1,4]]]]
[[2,[2,2]],[8,[8,1]]]
[2,9]
[1,[[[9,3],9],[[9,0],[0,7]]]]
[[[5,[7,4]],7],1]
[[[[4,2],2],6],[8,7]]
}

The final sum \verb{[[[[8,7],[7,7]],[[8,6],[7,7]]],[[[0,7],[6,6]],[8,7]]]} is
found after adding up the above snailfish numbers:\preformatted{  [[[0,[4,5]],[0,0]],[[[4,5],[2,6]],[9,5]]]
+ [7,[[[3,7],[4,3]],[[6,3],[8,8]]]]
= [[[[4,0],[5,4]],[[7,7],[6,0]]],[[8,[7,7]],[[7,9],[5,0]]]]

  [[[[4,0],[5,4]],[[7,7],[6,0]]],[[8,[7,7]],[[7,9],[5,0]]]]
+ [[2,[[0,8],[3,4]]],[[[6,7],1],[7,[1,6]]]]
= [[[[6,7],[6,7]],[[7,7],[0,7]]],[[[8,7],[7,7]],[[8,8],[8,0]]]]

  [[[[6,7],[6,7]],[[7,7],[0,7]]],[[[8,7],[7,7]],[[8,8],[8,0]]]]
+ [[[[2,4],7],[6,[0,5]]],[[[6,8],[2,8]],[[2,1],[4,5]]]]
= [[[[7,0],[7,7]],[[7,7],[7,8]]],[[[7,7],[8,8]],[[7,7],[8,7]]]]

  [[[[7,0],[7,7]],[[7,7],[7,8]]],[[[7,7],[8,8]],[[7,7],[8,7]]]]
+ [7,[5,[[3,8],[1,4]]]]
= [[[[7,7],[7,8]],[[9,5],[8,7]]],[[[6,8],[0,8]],[[9,9],[9,0]]]]

  [[[[7,7],[7,8]],[[9,5],[8,7]]],[[[6,8],[0,8]],[[9,9],[9,0]]]]
+ [[2,[2,2]],[8,[8,1]]]
= [[[[6,6],[6,6]],[[6,0],[6,7]]],[[[7,7],[8,9]],[8,[8,1]]]]

  [[[[6,6],[6,6]],[[6,0],[6,7]]],[[[7,7],[8,9]],[8,[8,1]]]]
+ [2,9]
= [[[[6,6],[7,7]],[[0,7],[7,7]]],[[[5,5],[5,6]],9]]

  [[[[6,6],[7,7]],[[0,7],[7,7]]],[[[5,5],[5,6]],9]]
+ [1,[[[9,3],9],[[9,0],[0,7]]]]
= [[[[7,8],[6,7]],[[6,8],[0,8]]],[[[7,7],[5,0]],[[5,5],[5,6]]]]

  [[[[7,8],[6,7]],[[6,8],[0,8]]],[[[7,7],[5,0]],[[5,5],[5,6]]]]
+ [[[5,[7,4]],7],1]
= [[[[7,7],[7,7]],[[8,7],[8,7]]],[[[7,0],[7,7]],9]]

  [[[[7,7],[7,7]],[[8,7],[8,7]]],[[[7,0],[7,7]],9]]
+ [[[[4,2],2],6],[8,7]]
= [[[[8,7],[7,7]],[[8,6],[7,7]]],[[[0,7],[6,6]],[8,7]]]
}

To check whether it\'s the right answer, the snailfish teacher only
checks the \emph{magnitude} of the final sum. The magnitude of a pair is 3
times the magnitude of its left element plus 2 times the magnitude of
its right element. The magnitude of a regular number is just that
number.

For example, the magnitude of \verb{[9,1]} is \code{3*9 + 2*1 = 29}; the magnitude
of \verb{[1,9]} is \code{3*1 + 2*9 = 21}. Magnitude calculations are recursive:
the magnitude of \verb{[[9,1],[1,9]]} is \code{3*29 + 2*21 = 129}.

Here are a few more magnitude examples:
\itemize{
\item \verb{[[1,2],[[3,4],5]]} becomes \code{143}.
\item \verb{[[[[0,7],4],[[7,8],[6,0]]],[8,1]]} becomes \code{1384}.
\item \verb{[[[[1,1],[2,2]],[3,3]],[4,4]]} becomes \code{445}.
\item \verb{[[[[3,0],[5,3]],[4,4]],[5,5]]} becomes \code{791}.
\item \verb{[[[[5,0],[7,4]],[5,5]],[6,6]]} becomes \code{1137}.
\item \verb{[[[[8,7],[7,7]],[[8,6],[7,7]]],[[[0,7],[6,6]],[8,7]]]} becomes
\code{3488}.
}

So, given this example homework assignment:\preformatted{[[[0,[5,8]],[[1,7],[9,6]]],[[4,[1,2]],[[1,4],2]]]
[[[5,[2,8]],4],[5,[[9,9],0]]]
[6,[[[6,2],[5,6]],[[7,6],[4,7]]]]
[[[6,[0,7]],[0,9]],[4,[9,[9,0]]]]
[[[7,[6,4]],[3,[1,3]]],[[[5,5],1],9]]
[[6,[[7,3],[3,2]]],[[[3,8],[5,7]],4]]
[[[[5,4],[7,7]],8],[[8,3],8]]
[[9,3],[[9,9],[6,[4,9]]]]
[[2,[[7,7],7]],[[5,8],[[9,3],[0,2]]]]
[[[[5,2],5],[8,[3,7]]],[[5,[7,5]],[4,4]]]
}

The final sum is:\preformatted{[[[[6,6],[7,6]],[[7,7],[7,0]]],[[[7,7],[7,7]],[[7,8],[9,9]]]]
}

The magnitude of this final sum is \code{4140}.

Add up all of the snailfish numbers from the homework assignment in the
order they appear. \emph{What is the magnitude of the final sum?}

\strong{Part Two}

You notice a second question on the back of the homework assignment:

What is the largest magnitude you can get from adding only two of the
snailfish numbers?

Note that snailfish addition is not
\href{https://en.wikipedia.org/wiki/Commutative_property}{commutative} - that
is, \code{x + y} and \code{y + x} can produce different results.

Again considering the last example homework assignment above:\preformatted{[[[0,[5,8]],[[1,7],[9,6]]],[[4,[1,2]],[[1,4],2]]]
[[[5,[2,8]],4],[5,[[9,9],0]]]
[6,[[[6,2],[5,6]],[[7,6],[4,7]]]]
[[[6,[0,7]],[0,9]],[4,[9,[9,0]]]]
[[[7,[6,4]],[3,[1,3]]],[[[5,5],1],9]]
[[6,[[7,3],[3,2]]],[[[3,8],[5,7]],4]]
[[[[5,4],[7,7]],8],[[8,3],8]]
[[9,3],[[9,9],[6,[4,9]]]]
[[2,[[7,7],7]],[[5,8],[[9,3],[0,2]]]]
[[[[5,2],5],[8,[3,7]]],[[5,[7,5]],[4,4]]]
}

The largest magnitude of the sum of any two snailfish numbers in this
list is \code{3993}. This is the magnitude of
\verb{[[2,[[7,7],7]],[[5,8],[[9,3],[0,2]]]]} +
\verb{[[[0,[5,8]],[[1,7],[9,6]]],[[4,[1,2]],[[1,4],2]]]}, which reduces to
\verb{[[[[7,8],[6,6]],[[6,0],[7,7]]],[[[7,8],[8,8]],[[7,9],[0,6]]]]}.

\emph{What is the largest magnitude of any sum of two different snailfish
numbers from the homework assignment?}
}
\examples{
f18a_snailfish_homework(example_data_18(5))
f18b_snailfish_homework2(example_data_18(5))
}
