% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day20.R
\name{day20}
\alias{day20}
\alias{f20_enhance_n_times}
\alias{example_data_20}
\title{Day 20: Trench Map}
\usage{
f20_enhance_n_times(x, n)

example_data_20(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f20a(x)} returns .... For Part Two,
\code{f20b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2021/day/20}{Trench Map}
}
\details{
\strong{Part One}

With the scanners fully deployed, you turn their attention to mapping
the floor of the ocean trench.

When you get back the image from the scanners, it seems to just be
random noise. Perhaps you can combine an image enhancement algorithm and
the input image (your puzzle input) to clean it up a little.

For example:\preformatted{..#.#..#####.#.#.#.###.##.....###.##.#..###.####..#####..#....#..#..##..##
#..######.###...####..#..#####..##..#.#####...##.#.#..#.##..#.#......#.###
.######.###.####...#.##.##..#..#..#####.....#.#....###..#.##......#.....#.
.#..#..##..#...##.######.####.####.#.#...#.......#..#.#.#...####.##.#.....
.#..#...##.#.##..#...##.#.##..###.#......#.#.......#.#.#.####.###.##...#..
...####.#..#..#.##.#....##..#.####....##...##..#...#......#.#.......#.....
..##..####..#...#.#.#...##..#.#..###..#####........#..####......#..#

#..#.
#....
##..#
..#..
..###
}

The first section is the \emph{image enhancement algorithm}. It is normally
given on a single line, but it has been wrapped to multiple lines in
this example for legibility. The second section is the \emph{input image}, a
two-dimensional grid of \emph{light pixels} (\verb{#}) and \emph{dark pixels} (\code{.}).

The image enhancement algorithm describes how to enhance an image by
\emph{simultaneously} converting all pixels in the input image into an output
image. Each pixel of the output image is determined by looking at a 3x3
square of pixels centered on the corresponding input image pixel. So, to
determine the value of the pixel at (5,10) in the output image, nine
pixels from the input image need to be considered: (4,9), (4,10),
(4,11), (5,9), (5,10), (5,11), (6,9), (6,10), and (6,11). These nine
input pixels are combined into a single binary number that is used as an
index in the \emph{image enhancement algorithm} string.

For example, to determine the output pixel that corresponds to the very
middle pixel of the input image, the nine pixels marked by \verb{[...]} would
need to be considered:\preformatted{# . . # .
#[. . .].
#[# . .]#
.[. # .].
. . # # #
}

Starting from the top-left and reading across each row, these pixels are
\code{...}, then \verb{#..}, then \code{.#.}; combining these forms \code{...#...#.}. By
turning dark pixels (\code{.}) into \code{0} and light pixels (\verb{#}) into \code{1}, the
binary number \code{000100010} can be formed, which is \code{34} in decimal.

The image enhancement algorithm string is exactly 512 characters long,
enough to match every possible 9-bit binary number. The first few
characters of the string (numbered starting from zero) are as follows:\preformatted{0         10        20        30  34    40        50        60        70
|         |         |         |   |     |         |         |         |
..#.#..#####.#.#.#.###.##.....###.##.#..###.####..#####..#....#..#..##..##
}

In the middle of this first group of characters, the character at index
34 can be found: \verb{#}. So, the output pixel in the center of the output
image should be \verb{#}, a \emph{light pixel}.

This process can then be repeated to calculate every pixel of the output
image.

Through advances in imaging technology, the images being operated on
here are \emph{infinite} in size. \emph{Every} pixel of the infinite output image
needs to be calculated exactly based on the relevant pixels of the input
image. The small input image you have is only a small region of the
actual infinite input image; the rest of the input image consists of
dark pixels (\code{.}). For the purposes of the example, to save on space,
only a portion of the infinite-sized input and output images will be
shown.

The starting input image, therefore, looks something like this, with
more dark pixels (\code{.}) extending forever in every direction not shown
here:\preformatted{...............
...............
...............
...............
...............
.....#..#......
.....#.........
.....##..#.....
.......#.......
.......###.....
...............
...............
...............
...............
...............
}

By applying the image enhancement algorithm to every pixel
simultaneously, the following output image can be obtained:\preformatted{...............
...............
...............
...............
.....##.##.....
....#..#.#.....
....##.#..#....
....####..#....
.....#..##.....
......##..#....
.......#.#.....
...............
...............
...............
...............
}

Through further advances in imaging technology, the above output image
can also be used as an input image! This allows it to be enhanced \emph{a
second time}:\preformatted{...............
...............
...............
..........#....
....#..#.#.....
...#.#...###...
...#...##.#....
...#.....#.#...
....#.#####....
.....#.#####...
......##.##....
.......###.....
...............
...............
...............
}

Truly incredible - now the small details are really starting to come
through. After enhancing the original input image twice, \code{35} pixels are
lit.

Start with the original input image and apply the image enhancement
algorithm twice, being careful to account for the infinite size of the
images. \emph{How many pixels are lit in the resulting image?}

\strong{Part Two}

You still can\'t quite make out the details in the image. Maybe you just
didn\'t
\href{https://en.wikipedia.org/wiki/Kernel_(image_processing)}{enhance} it
[enough]{title="Yeah, that's definitely the problem."}.

If you enhance the starting input image in the above example a total of
\emph{50} times, \code{3351} pixels are lit in the final output image.

Start again with the original input image and apply the image
enhancement algorithm 50 times. \emph{How many pixels are lit in the
resulting image?}
}
\examples{
f20_enhance_n_times(example_data_20(), 2)
f20b()
}
