% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day15.R
\name{dijkstra}
\alias{dijkstra}
\title{Compute least cost path using Dijkstra's algorithm}
\usage{
dijkstra(cost)
}
\arguments{
\item{cost}{cost matrix}
}
\description{
Compute least cost path using Dijkstra's algorithm
}
\examples{
file <- system.file("extdata", "day15.txt", package = "adventofcode21")
cost <- read_mat(file)
dijkstra(cost)

}
