% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day08.R
\name{easy_finder}
\alias{easy_finder}
\title{Part 1}
\usage{
easy_finder(file)
}
\arguments{
\item{file}{in file}
}
\description{
Part 1
}
\examples{
file <- system.file("extdata", "day8.txt", package = "adventofcode21")
easy_finder(file) # 303
decoded_sum(file) # 961734

}
