% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day20.R
\name{enhance_image}
\alias{enhance_image}
\title{image 'enhancement'}
\usage{
enhance_image(file, times)
}
\arguments{
\item{file}{input file}

\item{times}{number of times to enhance}
}
\description{
image 'enhancement'
}
\examples{
\dontrun{
file <- system.file("extdata", "day20.txt", package = "adventofcode21")
enhance_image(file, times = 2)
}

}
