% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f_filter}
\alias{f_filter}
\title{Pipe friendly version of \code{Filter()}}
\usage{
f_filter(x, f)
}
\arguments{
\item{x, f}{arguments for \code{Filter()}}
}
\value{
\code{Filter(f, x)}
}
\description{
Pipe friendly version of \code{Filter()}
}
