% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f_map2}
\alias{f_map2}
\title{Pipe friendly version of \code{Map()}}
\usage{
f_map2(x, y, f, ...)
}
\arguments{
\item{x, y, f, ...}{arguments for \code{Map()}}
}
\value{
\code{Map(f, x, y, ...)}
}
\description{
Pipe friendly version of \code{Map()}
}
