% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f_reduce}
\alias{f_reduce}
\title{Pipe friendly version of \code{Reduce()}}
\usage{
f_reduce(x, f, ...)
}
\arguments{
\item{x, f, ...}{arguments for \code{Reduce()}}
}
\value{
\code{Reduce(f, x, ...)}
}
\description{
Pipe friendly version of \code{Reduce()}
}
