% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day17.R
\name{find_maxima}
\alias{find_maxima}
\title{weird physics thing}
\usage{
find_maxima(goal_x, goal_y)
}
\arguments{
\item{goal_x}{min and max of x goal region}

\item{goal_y}{min and max of y goal region}
}
\description{
weird physics thing
}
\examples{
goal_x <- c(137, 171)
goal_y <- c(-98, -73)
d <- find_maxima(goal_x, goal_y)
max(d$ymax, na.rm = TRUE) # 4753
sum(!is.na(d$ymax)) # 1546

}
