% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day12.R
\name{find_twofer_paths}
\alias{find_twofer_paths}
\title{part 2: find all paths, allowing one "twofer"}
\usage{
find_twofer_paths(file)
}
\arguments{
\item{file}{input file}
}
\description{
part 2: find all paths, allowing one "twofer"
}
