% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day18.R
\name{fishy_mag}
\alias{fishy_mag}
\title{get fish math magnitude}
\usage{
fishy_mag(file)
}
\arguments{
\item{file}{input file}
}
\description{
get fish math magnitude
}
\examples{
\dontrun{
file <- system.file("extdata", "day18.txt", package = "adventofcode21")
fishy_mag(file)
}

}
