% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{gamma_epsilon_prod}
\alias{gamma_epsilon_prod}
\title{Part 1}
\usage{
gamma_epsilon_prod(f)
}
\arguments{
\item{f}{input file}
}
\description{
Part 1
}
\examples{
f <- system.file("extdata", "day3.txt", package = "adventofcode21")
gamma_epsilon_prod(f) # 2003336

}
