% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{get_aimed_result}
\alias{get_aimed_result}
\title{Get aimed result}
\usage{
get_aimed_result(instructions)
}
\arguments{
\item{instructions}{character vector}
}
\description{
Get aimed result
}
\examples{
f <- system.file("extdata", "day2.txt", package = "adventofcode21")
instructions <- read.delim(f, header = FALSE)[[1]]
get_aimed_result(instructions) # 1947878632

}
