% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day06.R
\name{grow}
\alias{grow}
\title{grow a population}
\usage{
grow(init, t)
}
\arguments{
\item{init}{initial age vector}

\item{t}{number of time steps}
}
\description{
grow a population
}
\examples{
init <- scan(system.file("extdata", "day6.txt", package = "adventofcode21"),
             sep = ",")
grow(init, t = 80)
grow(init, t = 256)

}
