% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day07.R
\name{least_cost}
\alias{least_cost}
\title{Moving crabbies on the cheap}
\usage{
least_cost(init)
}
\arguments{
\item{init}{initial positions}
}
\description{
Moving crabbies on the cheap
}
\examples{
init <- scan(system.file("extdata", "day7.txt", package = "adventofcode21"),
             sep = ",")
least_cost(init) # 344138
fancy_least_cost(init) # 94862124

}
