% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day09.R
\name{low_points}
\alias{low_points}
\title{find low points}
\usage{
low_points(file)
}
\arguments{
\item{file}{input}
}
\description{
find low points
}
\examples{
\dontrun{
file <- system.file("extdata", "day9.txt", package = "adventofcode21")
m <- read_mat(file)
mins <- low_points(file)

# part 1
sum(c(m)[which(as.logical(mins))] + 1) #560

# part 2
prod_biggest_basins(file) # 959136
}



}
