% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mod1}
\alias{mod1}
\title{\code{x \%\% m} but with a value between (0, m]}
\usage{
mod1(x, m)
}
\arguments{
\item{x}{number to divide}

\item{m}{number to divide with}
}
\value{
the remainder of \code{x \%\% m} but a remainder of 0 is
replaced with \code{m}
}
\description{
\code{x \%\% m} but with a value between (0, m]
}
\details{
Inspired by Julia's \code{mod1()}
https://docs.julialang.org/en/v1/base/math/#Base.mod1
}
