% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{move}
\alias{move}
\title{helper function - a}
\usage{
move(pos, dir, amt)
}
\arguments{
\item{pos}{vector of current position (horizonal, depth)}

\item{dir}{character. direction to move.}

\item{amt}{integer. amount to move}
}
\description{
helper function - a
}
