% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{move_b}
\alias{move_b}
\title{helper function - b}
\usage{
move_b(posb, dir, amt)
}
\arguments{
\item{posb}{vector of current position (horizonal, depth, aim)}

\item{dir}{character. direction to move.}

\item{amt}{integer. amount to move}
}
\description{
helper function - b
}
