% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{o2_co2_prod}
\alias{o2_co2_prod}
\title{Part 2}
\usage{
o2_co2_prod(f)
}
\arguments{
\item{f}{input file}
}
\description{
Part 2
}
\examples{
f <- system.file("extdata", "day3.txt", package = "adventofcode21")
o2_co2_prod(f) # 1877139
}
