% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day04.R
\name{pick_board}
\alias{pick_board}
\title{part 1}
\usage{
pick_board(f, goal = c("win", "lose"))
}
\arguments{
\item{f}{input file}

\item{goal}{win or lose}
}
\description{
part 1
}
\examples{
f <- system.file("extdata", "day4.txt", package = "adventofcode21")
pick_board(f, goal = "win") # 63552
pick_board(f, goal = "lose") # 9020

}
