% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day14.R
\name{polymerize}
\alias{polymerize}
\title{polymer fun}
\usage{
polymerize(file, steps)
}
\arguments{
\item{file}{input file}

\item{steps}{number of steps to run}
}
\description{
polymer fun
}
\examples{
file <- system.file("extdata", "day14.txt", package = "adventofcode21")
polymerize(file, 10)
polymerize(file, 40)

}
