% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day09.R
\name{prod_biggest_basins}
\alias{prod_biggest_basins}
\title{Find product of sizes for three largest basins}
\usage{
prod_biggest_basins(file)
}
\arguments{
\item{file}{input file}
}
\description{
Find product of sizes for three largest basins
}
