% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day16.R
\name{read_packets}
\alias{read_packets}
\title{best AoC ever}
\usage{
read_packets(bits, versions = NULL)
}
\arguments{
\item{bits}{input bits}

\item{versions}{initializes to NULL but changes when called recursively}
}
\description{
best AoC ever
}
\examples{
out <- system.file("extdata", "day16.txt", package = "adventofcode21") \%>\%
  readLines() \%>\%
  hex2bits() \%>\%
  read_packets()
sum(out$versions)
out$value # 539051801941

}
