% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{regex_sum}
\alias{regex_sum}
\title{Extract numbers from character vector matching a regex, then sum}
\usage{
regex_sum(pattern, x)
}
\arguments{
\item{pattern}{regular expression}

\item{x}{character vector}
}
\description{
Extract numbers from character vector matching a regex, then sum
}
\examples{
f <- system.file("extdata", "day2.txt", package = "adventofcode21")
instructions <- read.delim(f, header = FALSE)[[1]]
x_position <- regex_sum("forward", instructions)
y_positive <- regex_sum("down", instructions)
y_negative <- regex_sum("up", instructions)
y_position <- y_positive - y_negative
x_position * y_position # 1938402
}
