% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day01.R
\name{sliding_sum}
\alias{sliding_sum}
\title{Compute a sliding sum of a sequence}
\usage{
sliding_sum(x, width)
}
\arguments{
\item{x}{a sequence}

\item{width}{the width of the sliding sum}
}
\description{
Compute a sliding sum of a sequence
}
\examples{
f <- system.file("extdata", "day1.txt", package = "adventofcode21")
x <- read.delim(f, header = FALSE)[[1]]
lag_sums <- sliding_sum(x, width = 3)
count_increased(lag_sums) # 1395

}
