% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day19.R
\name{transmogrify_scanners}
\alias{transmogrify_scanners}
\title{Scanner flippin}
\usage{
transmogrify_scanners(file)
}
\arguments{
\item{input}{file}
}
\description{
Scanner flippin
}
\examples{
file <- system.file("extdata", "day19.txt", package = "adventofcode21")
transmogrify_scanners(file)

}
