% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day21.R
\name{day21}
\alias{day21}
\alias{f21a_play_deterministic_game}
\alias{f21b_play_quantum_game}
\alias{example_data_21}
\title{Day 21: Dirac Dice}
\usage{
f21a_play_deterministic_game(x)

f21b_play_quantum_game(x)

example_data_21(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f21a_play_deterministic_game(x)} returns the product
of the losing player's score and the total number of rolls. For Part Two,
\code{f21b_play_quantum_game(x)} returns the total win counts for each player.
}
\description{
\href{https://adventofcode.com/2021/day/21}{Dirac Dice}
}
\details{
\strong{Part One}

There\'s not much to do as you slowly descend to the bottom of the
ocean. The submarine computer [challenges you to a nice
game]{title="A STRANGE GAME."} of \emph{Dirac Dice}.

This game consists of a single
\href{https://en.wikipedia.org/wiki/Dice}{die}, two
\href{https://en.wikipedia.org/wiki/Glossary_of_board_games#piece}{pawns},
and a game board with a circular track containing ten spaces marked \code{1}
through \code{10} clockwise. Each player\'s \emph{starting space} is chosen
randomly (your puzzle input). Player 1 goes first.

Players take turns moving. On each player\'s turn, the player rolls the
die \emph{three times} and adds up the results. Then, the player moves their
pawn that many times \emph{forward} around the track (that is, moving
clockwise on spaces in order of increasing value, wrapping back around
to \code{1} after \code{10}). So, if a player is on space \code{7} and they roll \code{2},
\code{2}, and \code{1}, they would move forward 5 times, to spaces \code{8}, \code{9}, \code{10},
\code{1}, and finally stopping on \code{2}.

After each player moves, they increase their \emph{score} by the value of the
space their pawn stopped on. Players\' scores start at \code{0}. So, if the
first player starts on space \code{7} and rolls a total of \code{5}, they would
stop on space \code{2} and add \code{2} to their score (for a total score of \code{2}).
The game immediately ends as a win for any player whose score reaches
\emph{at least \code{1000}}.

Since the first game is a practice game, the submarine opens a
compartment labeled \emph{deterministic dice} and a 100-sided die falls out.
This die always rolls \code{1} first, then \code{2}, then \code{3}, and so on up to
\code{100}, after which it starts over at \code{1} again. Play using this die.

For example, given these starting positions:\preformatted{Player 1 starting position: 4
Player 2 starting position: 8
}

This is how the game would go:
\itemize{
\item Player 1 rolls \code{1}+\code{2}+\code{3} and moves to space \code{10} for a total score
of \code{10}.
\item Player 2 rolls \code{4}+\code{5}+\code{6} and moves to space \code{3} for a total score
of \code{3}.
\item Player 1 rolls \code{7}+\code{8}+\code{9} and moves to space \code{4} for a total score
of \code{14}.
\item Player 2 rolls \code{10}+\code{11}+\code{12} and moves to space \code{6} for a total
score of \code{9}.
\item Player 1 rolls \code{13}+\code{14}+\code{15} and moves to space \code{6} for a total
score of \code{20}.
\item Player 2 rolls \code{16}+\code{17}+\code{18} and moves to space \code{7} for a total
score of \code{16}.
\item Player 1 rolls \code{19}+\code{20}+\code{21} and moves to space \code{6} for a total
score of \code{26}.
\item Player 2 rolls \code{22}+\code{23}+\code{24} and moves to space \code{6} for a total
score of \code{22}.
}

\...after many turns\...
\itemize{
\item Player 2 rolls \code{82}+\code{83}+\code{84} and moves to space \code{6} for a total
score of \code{742}.
\item Player 1 rolls \code{85}+\code{86}+\code{87} and moves to space \code{4} for a total
score of \code{990}.
\item Player 2 rolls \code{88}+\code{89}+\code{90} and moves to space \code{3} for a total
score of \code{745}.
\item Player 1 rolls \code{91}+\code{92}+\code{93} and moves to space \code{10} for a final
score, \code{1000}.
}

Since player 1 has at least \code{1000} points, player 1 wins and the game
ends. At this point, the losing player had \code{745} points and the die had
been rolled a total of \code{993} times; \code{745 * 993 = 739785}.

Play a practice game using the deterministic 100-sided die. The moment
either player wins, \emph{what do you get if you multiply the score of the
losing player by the number of times the die was rolled during the
game?}

\strong{Part Two}

Now that you\'re warmed up, it\'s time to play the real game.

A second compartment opens, this time labeled \emph{Dirac dice}. Out of it
falls a single three-sided die.

As you experiment with the die, you feel a little strange. An
informational brochure in the compartment explains that this is a
\emph{quantum die}: when you roll it, the universe \emph{splits into multiple
copies}, one copy for each possible outcome of the die. In this case,
rolling the die always splits the universe into \emph{three copies}: one
where the outcome of the roll was \code{1}, one where it was \code{2}, and one
where it was \code{3}.

The game is played the same as before, although to prevent things from
getting too far out of hand, the game now ends when either player\'s
score reaches at least \code{21}.

Using the same starting positions as in the example above, player 1 wins
in \code{444356092776315} universes, while player 2 merely wins in
\code{341960390180808} universes.

Using your given starting positions, determine every possible outcome.
\emph{Find the player that wins in more universes; in how many universes does
that player win?}
}
\examples{
f21a_play_deterministic_game(example_data_21())
}
