% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day11.R
\name{find_all_flash}
\alias{find_all_flash}
\title{part 2: find when all things flash}
\usage{
find_all_flash(file)
}
\arguments{
\item{file}{input file}
}
\description{
part 2: find when all things flash
}
\examples{
\dontrun{
file <- system.file("extdata", "day11.txt", package = "adventofcode21")
find_all_flash(file)
}

}
