% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{pick_overlap}
\alias{pick_overlap}
\title{part 1}
\usage{
pick_overlap(f, allow_diagonals = TRUE)
}
\arguments{
\item{f}{input file}

\item{allow_diagonals}{whether to permit diagonal lines}
}
\description{
part 1
}
\examples{
f <- system.file("extdata", "day5.txt", package = "adventofcode21")
pick_overlap(f, allow_diagonals = FALSE)
pick_overlap(f, allow_diagonals = TRUE)

}
